requirements.txt
asgiref
boto3
botocore
certifi
cffi
charset-normalizer
cryptography
defusedxml
Django
django-allauth
django-storages
gunicorn
idna
jmespath
oauthlib
psycopg2
pycparser
PyJWT
python-dateutil
python-dotenv
python3-openid
requests
requests-oauthlib
s3transfer
six
social-auth-core
sqlparse
typing_extensions
tzdata
urllib3